fix malformed imports and extract handleError from main
Standardize end of file/function format
