+++ main.js
@@ -51,16 +51,6 @@ let fetchedData = {
     spaces: false
 };
 
-// Utility function to handle errors and display a user-friendly message
-const handleError = (error, message) => {
-    console.error(message, error);
-    const itemList = document.getElementById('itemList');
-    itemList.innerHTML = `<div class="text-red-500 text-center col-span-full p-8 bg-red-100 rounded-lg">
-        <p class="font-bold">Error loading items.</p>
-        <p>${message}</p>
-    </div>`;
-};
-
 //
 // SECTION 2: DATA FETCHING
 //
